refactor: 예산 수정 api PUT 방식으로 변경

+++ src/main/java/com/moa/repository/BudgetRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface BudgetRepository extends JpaRepository<Budget,Integer> {
+public interface BudgetRepository extends JpaRepository<Budget, Long> {
     @Query("SELECT b FROM Budget b WHERE b.user.userId = :userId AND b.startDate <= :date AND (b.endDate IS NULL OR b.endDate > :date)")
     List<Budget> findBudgetsByDate(LocalDate date, Long userId);
